shaders working!

+++ Mesh.cpp
@@ -274,7 +274,7 @@ Mesh::Mesh(const char* objFile)
 	//
 	// *************************************
 
-	CreateBuffers(verts.data(), verts.size(), indices.data(), indices.size());
+	CreateBuffers(verts.data(), (int)verts.size(), indices.data(), (int)indices.size());
 }
 
 Mesh::~Mesh() {
